Remove unneeded `PropTypes` from `LabelInput` and convert it to TS
